requirements.txt
python-telegram-bot==20.3
sentence-transformers==2.2.2
faiss-cpu==1.9.0
openai==1.52.1
numpy==1.23.5
